updated vcard
added v3 contact card that works on my phone, also rearranged some
of the main page to have hobbies in an order closer to what I want
to show and not based on what mummy thought was relevant to the
interview.

Also updated tupfile to use quote tags in html and leave apostraphe
unchanged.