coloursort now uses grid layout and new movement
the new movement algorithm that was discussed with ciaran is now
implemented, it should now do the least number of splits necessary and
allow levels with several shorter empties to be completable.

The gameboard now uses a grid layout, the tubes span over a number of
rows based on their capacity so each row basically represents a ball
height, this means tubes of different sizes can be packed together in
irregular fashion to get more dense arrangement. There is also a bit of
javascript code needed to help the css get the right heights but it just
measures the number of rows actually used in the grid so it is very
lightweight.