improved settings form
number inputs now are only 7 characters wide and number aligns to the
right to be clearer that it is intended to be part of the sentence.

Also explicitly go to index.html now after discovering that if the
current url had a # in it it wouldn't bring you to a new level unless
the parameters had been changed. Ideally I'd like to get a version that
links directly to the controls to help automate scrolling after starting
a new game but will need to think about how to actually do that given
this new info about linking to any id on the same page.

Also removed console log in the onload of the script that was left there by
mistake last commit