changed deployment strategy
apparently I was referencing some 3rd party code that tried to commit
all the files to the repo which is the exact thing I'm trying to avoid
by doing a build action. It should now hopefully compile the dist folder
as a build artifact and push that to pages.

And I know that theoretically I should make a new branch and test this
on that branch instead of master but I really can't be bothered this is
my personal website.