added solvability check
colour sort game now runs a solvability check in a second thread
as well as laying the ground work to add more functionality in seperate
worker thread,

I will need to re-organize the file structures to improve imports,
right now the worker thread imports all the UI code because it needs
the serialization function in the Tube class. I started a bit by
putting a commonHepers but even that should be split into smaller more
dedicated chunks. (worker helper class and callback should in theory
be in differnet modules as they will never be needed by the same side
although this may be overkill as it is tiny amount of code)