I'm (nearly) done with coloursort
I support arbitrary lists of empties capacities, "interesting twist" is now a preset revealing the hidden option. LevelCode is now properly a hidden setting that is revealed when set.

I just want to add a tiny bit of code to append a link to the list of presets to link to the current level